190 move TimeStampMapper to core
