+++ netKmm/src/commonMain/kotlin/uk/co/sentinelweb/cuer/net/youtube/videos/mapper/YoutubeChannelDomainMapper.kt
@@ -1,8 +1,8 @@
 package uk.co.sentinelweb.cuer.net.youtube.videos.mapper
 
+import uk.co.sentinelweb.cuer.core.mappers.TimeStampMapper
 import uk.co.sentinelweb.cuer.domain.ChannelDomain
 import uk.co.sentinelweb.cuer.domain.PlatformDomain
-import uk.co.sentinelweb.cuer.net.mappers.TimeStampMapper
 import uk.co.sentinelweb.cuer.net.youtube.videos.dto.YoutubeChannelsDto
 
 internal class YoutubeChannelDomainMapper(
